+++ src/main/java/com/plamaka/hotel_reservation_api/entity/Reservation.java
@@ -1,5 +1,6 @@
 package com.plamaka.hotel_reservation_api.entity;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -45,9 +46,9 @@ public class Reservation {
 	@Enumerated(EnumType.STRING)
 	private PaymentMethod paymentMethod;
 	
-	private Double depositAmount;
+	private BigDecimal depositAmount;
 	
-	private Double totalAmount;
+	private BigDecimal totalAmount;
 	
 
 	public Reservation() {
@@ -62,11 +63,11 @@ public class Reservation {
 		this.paymentMethod = paymentMethod;
 	}
 
-	public Double getTotalAmount() {
+	public BigDecimal getTotalAmount() {
 		return totalAmount;
 	}
 	
-	public void setTotalAmount(Double totalAmount) {
+	public void setTotalAmount(BigDecimal totalAmount) {
 		this.totalAmount = totalAmount;
 
 	}
@@ -127,11 +128,11 @@ public class Reservation {
 		this.paymentMethod = paymentMethod;
 	}
 
-	public Double getDepositAmount() {
+	public BigDecimal getDepositAmount() {
 		return depositAmount;
 	}
 
-	public void setDepositAmount(Double depositAmount) {
+	public void setDepositAmount(BigDecimal depositAmount) {
 		this.depositAmount = depositAmount;
 	}
